finagent/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finagent"
version = "0.1.0"
description = "Finance news agent for collecting and analyzing financial data"
authors = [{name = "Modugu Rohith Reddy"}]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.31.0",
    "openai>=1.1.1",
    "python-dotenv>=1.0.0",
    "pluggy>=1.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/rohithreddy1095/agents"
Issues = "https://github.com/rohithreddy1095/agents/issues"

[project.scripts]
finagent = "finagent.cli:cli"

[tool.setuptools.packages.find]
include = ["finagent*"]